+++ extensions/u301-url-shortener/CHANGELOG.md
@@ -1,5 +1,8 @@
 # U301 Url Shortener Changelog
 
+## [Custom Domain Support] - 2024-06-21
+- Support custom domain
+
 ## [Add API Key] - 2024-06-10
 - Support API Key